chore: add confirm modal

+++ src/modules/rec-services/history/helper.tsx
@@ -1,6 +1,6 @@
 import { getMultiSelectedItems } from '$components/RecGrid/rec-grid-state'
 import { isHistory } from '$define'
-import { antMessage } from '$modules/antd'
+import { antMessage, antModal } from '$modules/antd'
 import { normalizeCardData } from '$modules/filter/normalize'
 import { handleRequestError } from '$request'
 import { HistoryApiService } from './api'
@@ -9,6 +9,12 @@ import type { RecSharedEmitter } from '$components/Recommends/rec.shared'
 export async function removeMultiSelectedHistoryItems(recSharedEmitter: RecSharedEmitter) {
   const selected = getMultiSelectedItems().filter((x) => isHistory(x))
   if (!selected?.length) return antMessage.warning('没有选中项!')
+
+  const confirm = await antModal.confirm({
+    title: '移除历史',
+    content: <>确定移除 {selected.length} 条历史记录?</>,
+  })
+  if (!confirm) return
 
   const kid = selected.map((x) => `${x.history.business}_${x.kid}`).join(',')
   const result = await HistoryApiService.delete(kid)
